Add an option to keep users logged in for longer.
Add new optional config option:
$wgOAuth2Client['configuration']['remember_me']

It controls whether the users should be logged in the way they would if they
checked the "Keep me logged in" checkbox in password-based authentication.
Basically, the cookies are then set a bit differently, as described in
https://www.mediawiki.org/wiki/Manual:$wgCookieExpiration

The default is false, so the behaviour from before this patch is
preserved.

This is useful, as without it in some setups the user will be logged in
only for a short period of time, which makes e.g. editing experience
pretty annoying.